MSUtil replaced by MSUtility with categories
