+++ MobileCenter/MobileCenter/Internals/Channel/MSLogManagerDefault.m
@@ -108,7 +108,7 @@ static char *const MSlogsDispatchQueue = "com.microsoft.azure.mobile.mobilecente
   id<MSChannel> channel = [self channelForPriority:priority];
 
   // Set common log info.
-  log.toffset = [NSNumber numberWithLongLong:[MSUtil nowInMilliseconds]];
+  log.toffset = [NSNumber numberWithLongLong:[MSUtility nowInMilliseconds]];
 
   // Only add device info in case the log doesn't have one. In case the log is restored after a crash or for crashes,
   // We don't want the device information to be updated but want the old one preserved.
