+++ MobileCenter/MobileCenter/Internals/Utils/MSUtility.m
@@ -1,0 +1,18 @@
+#import "MSUtility+Application.h"
+#import "MSUtility+Environment.h"
+#import "MSUtility+Date.h"
+
+@implementation MSUtility
+
+/**
+ * @discussion
+ * Workaround for exporting symbols from category object files. 
+ * See article https://medium.com/ios-os-x-development/categories-in-static-libraries-78e41f8ddb96#.aedfl1kl0
+ */
+__attribute__((used)) static void importCategories () {
+    id applicationCategorySymbol = MSUtilityApplicationCategory;
+    id environmentCategorySymbol = MSUtilityEnvironmentCategory;
+    id dateCategorySymbol        = MSUtilityDateCategory;
+}
+
+@end
