+++ MobileCenter/MobileCenter/MSMobileCenter.m
@@ -167,7 +167,7 @@ static NSString *const kMSDefaultBaseUrl = @"https://in.mobile.azure.com";
        * If the loglevel hasn't been customized before and we are not running in an app store environment,
        * we set the default loglevel to MSLogLevelWarning.
        */
-      if ((![MSLogger isUserDefinedLogLevel]) && ([MSUtil currentAppEnvironment] == MSEnvironmentOther)) {
+      if ((![MSLogger isUserDefinedLogLevel]) && ([MSUtility currentAppEnvironment] == MSEnvironmentOther)) {
         [MSMobileCenter setLogLevel:MSLogLevelWarning];
       }
       success = true;
